+++ src/meson.build
@@ -47,6 +47,7 @@ xsystem4 = [version_h,
             'dungeon/dgn.c',
             'dungeon/dtx.c',
             'dungeon/dungeon.c',
+            'dungeon/generator.c',
             'dungeon/map.c',
             'dungeon/mtrand43.c',
             'dungeon/polyobj.c',
